+++ CHANGES.rst
@@ -6,6 +6,13 @@ Unreleased
 
 -
 
+Release 0.2b12 (2014-11-17)
+---------------------------
+
+- Timeline axis displays start and end of timeState in bold.
+
+- Various visual updates on the timeline.
+
 Release 0.2b11 (2014-11-12)
 ---------------------------